moving faq data to common web folder
