+++ common/src/web/index.ts
@@ -1,0 +1,1 @@
+export { FAQData, faqData } from "./faq-data";
